# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /company/factories:
    get:
      description: Returns every known factory. Pass owner_company_ifric_id to filter to just the factories owned by that company (the same id returned as company_ifric_id by /company/owners/:id and /company/factories/:id/owner).
      operationId: CompanyController_getFactories
      parameters:
        - name: owner_company_ifric_id
          required: false
          in: query
          description: Filter to factories owned by this company
          schema:
            example: urn:ifric:ifx-eur-com-own-42ced491-b35d-41f7-9949-fcbb5fa4dcd9
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get all factories, optionally filtered by owner
      tags:
        - Company
    post:
      description: Creates a physical factory location and tags it to an owner company via owner_company_ifric_id.
      operationId: CompanyController_createFactory
      parameters: []
      requestBody:
        required: true
        description: Details for creating a factory
        content:
          application/json:
            schema:
              type: object
              required:
                - factory_id
                - owner_company_ifric_id
              properties:
                factory_id:
                  type: string
                  example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
                owner_company_ifric_id:
                  type: string
                  example: urn:ifric:ifx-eur-com-own-42ced491-b35d-41f7-9949-fcbb5fa4dcd9
                location_name:
                  type: string
                  example: Plant 1
                address_1:
                  type: string
                  example: 123 Main St
                city:
                  type: string
                  example: Berlin
                country:
                  type: string
                  example: Germany
                zip:
                  type: string
                  example: '10115'
                latitude:
                  type: number
                  example: 52.52
                longitude:
                  type: number
                  example: 13.405
                timezone:
                  type: string
                  example: Europe/Berlin
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      summary: Create a factory tagged to an owner company
      tags:
        - Company
  /company/factories/{id}:
    get:
      description: Returns the physical factory location for a factory id, the same object shape as /company/factory-locations/:id but looked up directly by factory id rather than via a product URN.
      operationId: CompanyController_getFactoryById
      parameters:
        - name: id
          required: true
          in: path
          description: Factory id, e.g. urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
          schema:
            example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get factory details for a factory id
      tags:
        - Company
    patch:
      description: Updates the location details of an existing factory.
      operationId: CompanyController_updateFactory
      parameters:
        - name: id
          required: true
          in: path
          description: Factory id
          schema:
            example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Update a factory
      tags:
        - Company
    delete:
      description: Deletes a factory. Fails with 409 if any company twin still references this factory_id.
      operationId: CompanyController_deleteFactory
      parameters:
        - name: id
          required: true
          in: path
          description: Factory id
          schema:
            example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Delete a factory
      tags:
        - Company
  /company/factories/{id}/owner:
    get:
      description: Resolves a factory id to its owner_company_ifric_id and returns the full owner company object, the same schema as /company/owners/:id and /company/products/:id.
      operationId: CompanyController_getFactoryOwner
      parameters:
        - name: id
          required: true
          in: path
          description: Factory id, e.g. urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
          schema:
            example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get the owner company for a factory id
      tags:
        - Company
  /company/factories/{id}/products:
    get:
      description: Returns every product URN whose factory location (per /company/factory-locations/:id) resolves to this factory id. Pass any of the returned URNs to /company/products/:id to get that product's manufacturer.
      operationId: CompanyController_getFactoryProducts
      parameters:
        - name: id
          required: true
          in: path
          description: Factory id, e.g. urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
          schema:
            example: urn:ifric:ifx-eur-loc-fac-bd063b72-8748-461f-888d-3ea75058f205
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Get all product URNs located at a factory id
      tags:
        - Company
  /company/company-asset:
    post:
      operationId: CompanyController_createCompanyAsset
      parameters: []
      requestBody:
        required: true
        description: Details for creating a company asset
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - company_ifric_id
              properties:
                type:
                  type: string
                  enum:
                    - asset
                    - gateway
                    - server
                  example: asset
                company_ifric_id:
                  type: string
                  example: IFRIC12345
                asset_ifric_id:
                  type: string
                  description: Required when type is "asset"
                  example: ASSET67890
                gateway_ifric_id:
                  type: string
                  description: Required when type is "gateway"
                  example: GATEWAY112233
                server_ifric_id:
                  type: string
                  description: Required when type is "server"
                  example: SERVER445566
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
  /company/create-access-group/{id}:
    post:
      operationId: CompanyController_createAccessGroup
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: Details for creating an access group
        content:
          application/json:
            schema:
              type: object
              properties:
                company_id:
                  type: string
                  example: COMPANY12345
                group_name:
                  type: string
                  example: Admin Group
                create:
                  type: boolean
                  example: true
                read:
                  type: boolean
                  example: true
                update:
                  type: boolean
                  example: false
                delete:
                  type: boolean
                  example: false
              required:
                - company_id
                - group_name
                - create
                - read
                - update
                - delete
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
  /company/create-company:
    post:
      operationId: CompanyController_createCompany
      parameters: []
      requestBody:
        required: true
        description: Details for creating a company
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                  example: Example Company Ltd.
                registration_number:
                  type: string
                  example: REG123456
                company_ifric_id:
                  type: string
                  example: IFRIC54321
                address_1:
                  type: string
                  example: 123 Main Street
                city:
                  type: string
                  example: New York
                country:
                  type: string
                  example: USA
                zip:
                  type: string
                  example: '10001'
                admin_name:
                  type: string
                  example: John Admin
                position:
                  type: string
                  example: CEO
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: strongpassword123
                company_size:
                  type: string
                  example: 100-500
                company_category_id:
                  type: number
                  example: 1
                company_category:
                  type: string
                  example: IT Services
                meta_data:
                  type: object
                  additionalProperties: true
                  example:
                    industry: Technology
                    revenue: 1M+
                company_domain:
                  type: string
                  example: example.com
              required:
                - company_name
                - registration_number
                - company_ifric_id
                - address_1
                - city
                - country
                - zip
                - admin_name
                - position
                - email
                - password
                - company_size
                - company_category
                - meta_data
                - company_domain
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
  /company/add-status-detail:
    post:
      operationId: CompanyController_addStatusDetail
      parameters: []
      requestBody:
        required: true
        description: Add a status for a company
        content:
          application/json:
            schema:
              type: object
              properties:
                company_id:
                  type: string
                  example: COMPANY_ID_123
                  description: The unique ID of the company
                status:
                  type: string
                  example: verified
                  description: The status to be set for the company (e.g., new, verify, verified, unverified)
              required:
                - company_id
                - status
      responses:
        '201':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
  /company/get-company-assets/{id}:
    get:
      operationId: CompanyController_getCompanyAssets
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
  /company/get-company-assets-by-asset/{assetId}:
    get:
      operationId: CompanyController_getCompanyAssetsbyAsset
      parameters:
        - name: assetId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
  /company/get-company-asset-by-assetid/{asset_ifric_id}:
    get:
      operationId: CompanyController_getCompanyAssetByAssetId
      parameters:
        - name: asset_ifric_id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
...
  /company/get-all-companies:
    get:
      operationId: CompanyController_getAllCompanies
      parameters: []
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      tags:
        - Company
...
info:
  title: Ifric Registry Service
  description: Open-source multi-tenant company/user/access-control/digital-twin registry service, split into Auth (session/identity), Company (company CRUD, access groups, physical assets), Product (product catalog, digital twins), Certificate, and Script (one-time seed data) controllers. All use a single bearer-JWT auth mechanism (POST /auth/login, refreshed via POST /auth/refresh). Certificate issuance and company_ifric_id minting integrate with an external open-source ICID-compatible service (see README.md).
  version: 0.1.0
  contact: {}
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
...